+++ CHANGELOG.rst
@@ -11,8 +11,25 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 `Unreleased`_
 -------------
 
-Nothing yet.
+Added
+~~~~~
 
+- Appveyor builds for compiled wheels for win32/win64 x86
+
+Bugfix
+~~~~~~
+
+- Fixes missing import statement in room.plot for 3D rooms (PR #286)
+- On win64, ``bss.fastmnmf`` would fail due to some singular matrix. 1) protect solve
+  with try/except and switch to pseudo-inverse if necessary, 2) change eps 1e-7 -> 1e-6
+
+`0.7.1`_ - 2022-11-11
+---------------------
+
+Bugfix
+~~~~~~
+
+- Fixed pypi upload for windows wheels
 
 `0.7.0`_ - 2022-11-10
 ---------------------
@@ -486,7 +503,8 @@ Changed
    ``pyroomacoustics.datasets.timit``
 
 
-.. _Unreleased: https://github.com/LCAV/pyroomacoustics/compare/v0.7.0...master
+.. _Unreleased: https://github.com/LCAV/pyroomacoustics/compare/v0.7.1...master
+.. _0.7.1: https://github.com/LCAV/pyroomacoustics/compare/v0.7.0...v0.7.1
 .. _0.7.0: https://github.com/LCAV/pyroomacoustics/compare/v0.6.0...v0.7.0
 .. _0.6.0: https://github.com/LCAV/pyroomacoustics/compare/v0.5.0...v0.6.0
 .. _0.5.0: https://github.com/LCAV/pyroomacoustics/compare/v0.4.3...v0.5.0